pyproject.toml
[project]
name = "family-book"
version = "2.0.0"
description = "FamilyBook 2.0 - Private Family Social Network with PWA and Guest Sandbox"
authors = [
    {name = "Mihhail Karasjov", email = "114910034+Mihhail327@users.noreply.github.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "bcrypt>=4.1.0",
    "pillow>=10.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "itsdangerous>=2.2.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic (>=1.18.4,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-playwright (>=0.7.2,<0.8.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
